pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot==20.8",
    "python-dotenv>=1.0.0",
    "flask>=3.1.1",
    "requests>=2.32.4",
    "telegram>=0.0.1",
]
